+++ main.ts
@@ -88,7 +88,7 @@ export default class AbstractFolderPlugin extends Plugin {
 			callback: () => {
 				new CreateAbstractChildModal(this.app, this.settings, (childName: string, childType: ChildFileType) => {
 					new ParentPickerModal(this.app, (parentFile) => {
-						createAbstractChildFile(this.app, this.settings, childName, parentFile, childType)
+						createAbstractChildFile(this.app, this.settings, childName, parentFile, childType, this.indexer)
 							.catch(console.error);
 					}).open();
 				}).open();